create AccountController; move GetAccounts to AccountController
